+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (#2664)
 - Bumped the credential version to 2 to support Posit Connect Cloud credentials. (#2684)
 
+### Added
+
+- Improved support for Posit Connect deployments hosted in Snowpark Container
+  Services. (#2687, #2691)
+
 ## [1.16.1]
 
 ### Fixed